adds char encoding in html example
IronPdf new renderer is based on Chrome. If you try open_me_in_chrome_browser.html in chrome you will find it does not render the Chinese properly. Chrome browser checks the first 60,000 or so characters to auto-detect encoding and thus a very long HTML string with special chars at the end (but not beginning) will fail the encoding detection in Chrome. Therefor it is best practice to specify encoding at the start of your html (<meta charset='utf-8'/>) to ensure it works in all browsers.

However, the upcoming IronPdf release (within a week or so) will attempt to detect the absence of explicit character encoding and will automatically prepend <meta charset='utf-8'/> to your html so this issue should go away